+++ packages/core/datetime-picker/src/components/DatePicker.tsx
@@ -7,7 +7,7 @@ import {
   createLocalizationProvider,
   LocalizationProvider,
 } from '@atlaskit/locale';
-import { borderRadius, layers } from '@atlaskit/theme/constants';
+import { borderRadius, layers, gridSize } from '@atlaskit/theme/constants';
 import { N20, B100 } from '@atlaskit/theme/colors';
 import { e200 } from '@atlaskit/theme/elevation';
 import {
@@ -445,6 +445,9 @@ class DatePicker extends React.Component<Props, State> {
       spacing,
       locale,
     } = this.props;
+    const BORDER_WIDTH = 2;
+    const ICON_PADDING = 2;
+
     const { value, view, isOpen, inputValue } = this.getSafeState();
     const dropDownIcon = appearance === 'subtle' || hideIcon ? null : icon;
     const { styles: selectStyles = {} } = selectProps;
@@ -494,6 +497,11 @@ class DatePicker extends React.Component<Props, State> {
               ...base,
               ...controlStyles,
               ...disabledStyle,
+            }),
+            indicatorsContainer: (base: CSSObject): CSSObject => ({
+              ...base,
+              paddingLeft: ICON_PADDING,
+              paddingRight: gridSize() - BORDER_WIDTH,
             }),
           })}
           placeholder={this.getPlaceholder()}