+++ packages/core/input/src/__tests__/unit/server-side-hydrate.js
@@ -6,18 +6,14 @@ import { ssr } from '@atlaskit/ssr';
 
 jest.spyOn(global.console, 'error').mockImplementation(() => {});
 
-beforeEach(() => {
-  jest.setTimeout(10000);
-});
-
 afterEach(() => {
   jest.resetAllMocks();
 });
-// TODO: https://ecosystem.atlassian.net/browse/AK-6450// https://product-fabric.atlassian.net/browse/BUILDTOOLS-282: SSR tests are still timing out in Landkid.
+// TODO: https://ecosystem.atlassian.net/browse/AK-6450
 test.skip('should ssr then hydrate input correctly', async () => {
   const [example] = await getExamplesFor('input');
   // $StringLitteral
-  const Example = await require(example.filePath).default; // eslint-disable-line import/no-dynamic-require
+  const Example = require(example.filePath).default; // eslint-disable-line import/no-dynamic-require
 
   const elem = document.createElement('div');
   elem.innerHTML = await ssr(example.filePath);