+++ packages/media/media-client/src/__tests__/unit/media-client-server-side-hydrate.tsx
@@ -4,16 +4,14 @@ import { getExamplesFor } from '@atlaskit/build-utils/getExamples';
 import { ssr, mockConsole } from '@atlaskit/ssr';
 
 const getConsoleMockCalls = mockConsole(console);
-beforeAll(() => {
-  jest.setTimeout(30000);
-});
+
 afterEach(() => {
   jest.resetAllMocks();
 });
-// https://product-fabric.atlassian.net/browse/BUILDTOOLS-282: SSR tests are still timing out in Landkid.
-test.skip('should ssr then hydrate media-client correctly', async () => {
+
+test('should ssr then hydrate media-client correctly', async () => {
   const [example] = await getExamplesFor('media-client');
-  const Example = await require(example.filePath).default;
+  const Example = require(example.filePath).default;
   const elem = document.createElement('div');
   elem.innerHTML = await ssr(example.filePath);