+++ packages/navigation/atlassian-switcher/src/components/switcher.tsx
@@ -52,6 +52,12 @@ export type SwitcherProps = {
    */
   disableHeadings?: boolean;
   appearance?: Appearance;
+  /**
+   * Links for experimental "Discover" section
+   * which is a variation of suggestedProductLinks and fixedLinks combined
+   */
+  isDiscoverSectionEnabled?: boolean;
+  discoverSectionLinks: SwitcherItemType[];
 };
 
 const getAnalyticsContext = (itemsCount: number) => ({
@@ -61,7 +67,7 @@ const getAnalyticsContext = (itemsCount: number) => ({
 });
 
 const getItemAnalyticsContext = (
-  index: number,
+  groupIndex: number,
   id: string | null,
   type: string,
   href: string,
@@ -69,7 +75,7 @@ const getItemAnalyticsContext = (
   extraAttributes?: { [key: string]: string },
 ) => ({
   ...analyticsAttributes({
-    groupItemIndex: index,
+    groupIndex,
     itemId: id,
     itemType: type,
     domain: urlToHostname(href),
@@ -130,8 +136,9 @@ export default class Switcher extends React.Component<SwitcherProps> {
       hasLoadedCritical,
       disableHeadings,
       appearance,
+      isDiscoverSectionEnabled,
+      discoverSectionLinks,
     } = this.props;
-
     /**
      * It is essential that switchToLinks reflects the order corresponding nav items
      * are rendered below in the 'Switch to' section.
@@ -144,7 +151,10 @@ export default class Switcher extends React.Component<SwitcherProps> {
     ];
 
     const itemsCount =
-      switchToLinks.length + recentLinks.length + customLinks.length;
+      switchToLinks.length +
+      recentLinks.length +
+      customLinks.length +
+      discoverSectionLinks.length;
 
     const firstContentArrived = Boolean(licensedProductLinks.length);
 
@@ -192,11 +202,11 @@ export default class Switcher extends React.Component<SwitcherProps> {
               )
             }
           >
-            {licensedProductLinks.map(item => (
+            {licensedProductLinks.map((item, groupIndex) => (
               <NavigationAnalyticsContext
                 key={item.key}
                 data={getItemAnalyticsContext(
-                  switchToLinks.indexOf(item),
+                  groupIndex,
                   item.key,
                   'product',
                   item.href,
@@ -217,32 +227,33 @@ export default class Switcher extends React.Component<SwitcherProps> {
                 </SwitcherItemWithDropdown>
               </NavigationAnalyticsContext>
             ))}
-            {suggestedProductLinks.map(item => (
-              <NavigationAnalyticsContext
-                key={item.key}
-                data={getItemAnalyticsContext(
-                  switchToLinks.indexOf(item),
-                  item.key,
-                  'try',
-                  item.href,
-                )}
-              >
-                <SwitcherThemedItemWithEvents
-                  icon={<item.Icon theme="product" />}
-                  onClick={this.triggerXFlow(item.key)}
+            {!isDiscoverSectionEnabled &&
+              suggestedProductLinks.map((item, groupIndex) => (
+                <NavigationAnalyticsContext
+                  key={item.key}
+                  data={getItemAnalyticsContext(
+                    groupIndex,
+                    item.key,
+                    'try',
+                    item.href,
+                  )}
                 >
-                  {item.label}
-                  <TryLozenge>
-                    <FormattedMessage {...messages.try} />
-                  </TryLozenge>
-                </SwitcherThemedItemWithEvents>
-              </NavigationAnalyticsContext>
-            ))}
-            {fixedLinks.map(item => (
+                  <SwitcherThemedItemWithEvents
+                    icon={<item.Icon theme="product" />}
+                    onClick={this.triggerXFlow(item.key)}
+                  >
+                    {item.label}
+                    <TryLozenge>
+                      <FormattedMessage {...messages.try} />
+                    </TryLozenge>
+                  </SwitcherThemedItemWithEvents>
+                </NavigationAnalyticsContext>
+              ))}
+            {fixedLinks.map((item, groupIndex) => (
               <NavigationAnalyticsContext
                 key={item.key}
                 data={getItemAnalyticsContext(
-                  switchToLinks.indexOf(item),
+                  groupIndex,
                   item.key,
                   'product',
                   item.href,
@@ -261,11 +272,11 @@ export default class Switcher extends React.Component<SwitcherProps> {
                 </SwitcherThemedItemWithEvents>
               </NavigationAnalyticsContext>
             ))}
-            {adminLinks.map(item => (
+            {adminLinks.map((item, groupIndex) => (
               <NavigationAnalyticsContext
                 key={item.key}
                 data={getItemAnalyticsContext(
-                  switchToLinks.indexOf(item),
+                  groupIndex,
                   item.key,
                   'admin',
                   item.href,
@@ -280,6 +291,65 @@ export default class Switcher extends React.Component<SwitcherProps> {
               </NavigationAnalyticsContext>
             ))}
           </Section>
+          )}
+          {isDiscoverSectionEnabled && (
+            <Section
+              sectionId="discover"
+              title={
+                disableHeadings ? null : (
+                  <FormattedMessage {...messages.discover} />
+                )
+              }
+            >
+              {suggestedProductLinks.map((item, groupIndex) => (
+                <NavigationAnalyticsContext
+                  key={item.key}
+                  data={getItemAnalyticsContext(
+                    groupIndex,
+                    item.key,
+                    'discover',
+                    item.href,
+                  )}
+                >
+                  <SwitcherThemedItemWithEvents
+                    icon={<item.Icon theme="recommendedProduct" />}
+                    description={item.description}
+                    onClick={this.triggerXFlow(item.key)}
+                  >
+                    {item.label}
+                    {groupIndex === 0 && (
+                      <TryLozenge isBold={false}>
+                        <FormattedMessage {...messages.try} />
+                      </TryLozenge>
+                    )}
+                  </SwitcherThemedItemWithEvents>
+                </NavigationAnalyticsContext>
+              ))}
+              {discoverSectionLinks.map((item, groupIndex) => (
+                <NavigationAnalyticsContext
+                  key={item.key}
+                  data={getItemAnalyticsContext(
+                    groupIndex,
+                    item.key,
+                    'discover-fixed-links',
+                    item.href,
+                  )}
+                >
+                  <SwitcherThemedItemWithEvents
+                    icon={<item.Icon theme="discover" />}
+                    href={item.href}
+                    onClick={
+                      item.key === 'discover-more'
+                        ? this.onDiscoverMoreClicked
+                        : noop
+                    }
+                  >
+                    {item.label}
+                  </SwitcherThemedItemWithEvents>
+                </NavigationAnalyticsContext>
+              ))}
+            </Section>
+          )}
           <Section
             sectionId="recent"
             title={
@@ -287,10 +357,15 @@ export default class Switcher extends React.Component<SwitcherProps> {
             }
           >
             {recentLinks.map(
-              ({ key, label, href, type, description, Icon }, idx) => (
+              ({ key, label, href, type, description, Icon }, groupIndex) => (
                 <NavigationAnalyticsContext
                   key={key}
-                  data={getItemAnalyticsContext(idx, type, 'recent', href)}
+                  data={getItemAnalyticsContext(
+                    groupIndex,
+                    type,
+                    'recent',
+                    href,
+                  )}
                 >
                   <SwitcherThemedItemWithEvents
                     icon={<Icon theme="recent" />}
@@ -310,12 +385,12 @@ export default class Switcher extends React.Component<SwitcherProps> {
             }
           >
             {customLinks.map(
-              ({ analyticsAttributes, label, href, Icon }, idx) => (
+              ({ analyticsAttributes, label, href, Icon }, groupIndex) => (
                 // todo: id in SwitcherItem should be consumed from custom link resolver
                 <NavigationAnalyticsContext
-                  key={idx + '.' + label}
+                  key={groupIndex + '.' + label}
                   data={getItemAnalyticsContext(
-                    idx,
+                    groupIndex,
                     null,
                     'customLink',
                     href,