+++ packages/navigation/atlassian-switcher/src/utils/links.tsx
@@ -69,22 +69,25 @@ export const getObjectTypeLabel = (type: string): React.ReactNode => {
 export const getFixedProductLinks = (params: {
   isDiscoverMoreForEveryoneEnabled: boolean;
 }): SwitcherItemType[] => {
-  return params.isDiscoverMoreForEveryoneEnabled
-    ? [
-        {
-          // The discover more link href is intentionally empty to prioritise the onDiscoverMoreClicked callback
-          key: 'discover-more',
-          label: <FormattedMessage {...messages.discoverMore} />,
-          Icon: createIcon(AddIcon, { size: 'medium' }),
-          href: '',
-        },
-      ]
-    : [];
+  return params.isDiscoverMoreForEveryoneEnabled ? [getDiscoverMoreLink()] : [];
+};
+
+const getDiscoverMoreLink = (
+  customIcon?: React.ComponentType<any>,
+): SwitcherItemType => {
+  const icon = customIcon || AddIcon;
+  return {
+    // The discover more link href is intentionally empty to prioritise the onDiscoverMoreClicked callback
+    key: 'discover-more',
+    label: <FormattedMessage {...messages.discoverMore} />,
+    Icon: createIcon(icon, { size: 'medium' }),
+    href: '',
+  };
 };
 
 type AvailableProductDetails = Pick<
   SwitcherItemType,
-  'label' | 'Icon' | 'href'
+  'label' | 'Icon' | 'href' | 'description'
 >;
 
 export const AVAILABLE_PRODUCT_DATA_MAP: {
@@ -99,6 +102,9 @@ export const AVAILABLE_PRODUCT_DATA_MAP: {
     label: 'Confluence',
     Icon: createIcon(ConfluenceIcon, { size: 'small' }),
     href: '/wiki',
+    description: (
+      <FormattedMessage {...messages.productDescriptionConfluence} />
+    ),
   },
   [WorklensProductType.JIRA_BUSINESS]: {
     label: 'Jira Core',
@@ -109,16 +115,23 @@ export const AVAILABLE_PRODUCT_DATA_MAP: {
     label: 'Jira Software',
     Icon: createIcon(JiraSoftwareIcon, { size: 'small' }),
     href: '/secure/BrowseProjects.jspa?selectedProjectType=software',
+    description: (
+      <FormattedMessage {...messages.productDescriptionJiraSoftware} />
+    ),
   },
   [WorklensProductType.JIRA_SERVICE_DESK]: {
     label: 'Jira Service Desk',
     Icon: createIcon(JiraServiceDeskIcon, { size: 'small' }),
     href: '/secure/BrowseProjects.jspa?selectedProjectType=service_desk',
+    description: (
+      <FormattedMessage {...messages.productDescriptionJiraServiceDesk} />
+    ),
   },
   [WorklensProductType.OPSGENIE]: {
     label: 'Opsgenie',
     Icon: createIcon(OpsGenieIcon, { size: 'small' }),
     href: 'https://app.opsgenie.com',
+    description: <FormattedMessage {...messages.productDescriptionOpsgenie} />,
   },
   [WorklensProductType.STATUSPAGE]: {
     label: 'Statuspage',
@@ -248,9 +261,10 @@ export const getAdministrationLinks = (
   isDiscoverMoreForEveryoneEnabled: boolean,
   isEmceeLinkEnabled: boolean,
   product?: Product,
+  isDiscoverSectionEnabled?: boolean,
 ): SwitcherItemType[] => {
   const adminBaseUrl = isAdmin ? `/admin` : '/trusted-admin';
-  const adminLinks = [
+  const adminLinks: SwitcherItemType[] = [
     {
       key: 'administration',
       label: <FormattedMessage {...messages.administration} />,
@@ -258,15 +272,16 @@ export const getAdministrationLinks = (
       href: adminBaseUrl,
     },
   ];
-  const emceeLink = product && BROWSE_APPS_URL[product];
-  if (isEmceeLinkEnabled && emceeLink) {
-    adminLinks.unshift({
-      key: 'browse-apps',
-      label: <FormattedMessage {...messages.browseApps} />,
-      Icon: createIcon(MarketplaceGlyph, { size: 'medium' }),
-      href: `${emceeLink}#!/discover?source=app_switcher`,
-    });
+
+  if (isDiscoverSectionEnabled) {
+    return adminLinks;
   }
+
+  const emceeLink = isEmceeLinkEnabled && getEmceeLink(product);
+  if (emceeLink) {
+    adminLinks.unshift(emceeLink);
+  }
+
   if (!isDiscoverMoreForEveryoneEnabled) {
     adminLinks.unshift({
       key: 'discover-applications',
@@ -278,11 +293,26 @@ export const getAdministrationLinks = (
   return adminLinks;
 };
 
+const getEmceeLink = (product?: Product): SwitcherItemType | undefined => {
+  const emceeLink = product && BROWSE_APPS_URL[product];
+
+  if (emceeLink) {
+    return {
+      key: 'browse-apps',
+      label: <FormattedMessage {...messages.browseApps} />,
+      Icon: createIcon(MarketplaceGlyph, { size: 'medium' }),
+      href: `${emceeLink}#!/discover?source=app_switcher`,
+    };
+  }
+};
+
 const PRODUCT_RECOMMENDATION_LIMIT = 2;
+const DISCOVER_PRODUCT_RECOMMENDATION_LIMIT = 3;
 
 export const getSuggestedProductLink = (
   provisionedProducts: ProvisionedProducts,
   productRecommendations: RecommendationsEngineResponse,
+  isDiscoverSectionEnabled?: boolean,
 ): SwitcherItemType[] => {
   return productRecommendations
     .filter(legacyProduct => {
@@ -296,8 +326,39 @@ export const getSuggestedProductLink = (
         ...AVAILABLE_PRODUCT_DATA_MAP[productKey],
       };
     })
-    .slice(0, PRODUCT_RECOMMENDATION_LIMIT);
+    .slice(
+      0,
+      isDiscoverSectionEnabled
+        ? DISCOVER_PRODUCT_RECOMMENDATION_LIMIT
+        : PRODUCT_RECOMMENDATION_LIMIT,
+    );
 };
+
+export function getDiscoverSectionLinks({
+  isDiscoverMoreForEveryoneEnabled,
+  isEmceeLinkEnabled,
+  product,
+}: {
+  isDiscoverMoreForEveryoneEnabled: boolean;
+  isEmceeLinkEnabled: boolean;
+  product?: Product;
+}) {
+  const discoverLinks: SwitcherItemType[] = [];
+  const discoverMoreLink =
+    isDiscoverMoreForEveryoneEnabled &&
+    getDiscoverMoreLink(DiscoverFilledGlyph);
+  const emceeLink = isEmceeLinkEnabled && getEmceeLink(product);
+
+  if (discoverMoreLink) {
+    discoverLinks.push(discoverMoreLink);
+  }
+
+  if (emceeLink) {
+    discoverLinks.push(emceeLink);
+  }
+
+  return discoverLinks;
+}
 
 export const getProvisionedProducts = (
   availableProducts: AvailableProductsResponse,