+++ packages/core/onboarding/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/onboarding
 
+## 9.0.4
+
+### Patch Changes
+
+- [patch][f86839ca4e](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/f86839ca4e):
+
+  @atlaskit/portal had an issue in IE11 and this is fixed in 3.1.2
+
 ## 9.0.3
 
 ### Patch Changes